--- NAPARI.md
@@ -1,11 +1,0 @@
-# Napari
-
-## To set up napari proceed with these steps
-
-download and set up miniconda at https://www.anaconda.com/download
-
-open its terminal and create a napari enviroment by running these commands:
-
-'conda create -y -n napari-env -c conda-forge python=3.9'
-
-python -m pip install "napari[all]"'
