+++ peregrin/peregrin.egg-info/PKG-INFO
@@ -1,0 +1,9 @@
+Metadata-Version: 2.1
+Name: peregrin
+Version: 0.0.1
+Summary: Demo
+Home-page: https://github.com/BranislavModriansky/Peregrin/tree/main
+Author: Branislav Modriansky
+Author-email: modrinaskybranislav@gmail.com
+License: MIT
+Requires-Dist: os
