+++ peregrin/peregrin.egg-info/SOURCES.txt
@@ -1,0 +1,9 @@
+setup.py
+peregrin/__init__.py
+peregrin/scripts.py
+peregrin.egg-info/PKG-INFO
+peregrin.egg-info/SOURCES.txt
+peregrin.egg-info/dependency_links.txt
+peregrin.egg-info/requires.txt
+peregrin.egg-info/top_level.txt
+tests/test_scripts.py
