+++ peregrin/project.toml
@@ -1,0 +1,14 @@
+[build-system]
+requires = ["setuptools>=40.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "peregrin"
+version = "0.0.1"
+authors = [{name = "Branislav Modriansky", email = "modrianskybranislav@gmail.com"}]
+classifiers = [
+  "Development Status :: 3 - Alpha",
+  "Programming Language :: Python",
+  "Intended Audience :: Data Analysts",
+  "License :: OSI Approved :: MIT License",
+]
